Remove confusion as to the party list working
